requirements.txt
accelerate
absl-py
bitsandbytes
datasets
chardet
goatools
evaluate
gpytorch
matplotlib
numpy
pandas
peft
pynvml
scikit-learn
seaborn
tqdm
transformers[torch]
torchvision
trl
umap-learn[plot]
wandb
